+++ lib/fmtc_advanced.dart
@@ -7,7 +7,7 @@ library fmtc_advanced;
 
 export 'flutter_map_tile_caching.dart';
 export 'src/bulk_download/downloader.dart';
-export 'src/bulk_download/progress_management.dart';
+export 'src/bulk_download/internal_timing_progress_management.dart';
 export 'src/bulk_download/tile_loops.dart';
 export 'src/bulk_download/tile_progress.dart';
 export 'src/internal/exts.dart';