+++ lib/src/store/statistics.dart
@@ -198,7 +198,7 @@ class StoreStats {
   ///
   /// For asynchronous version, see [cacheMissesAsync].
   ///
-  /// If using [noCache], this will always return 0.
+  /// Includes tiles that needed updating. If using [noCache], this will always return 0.
   int get cacheMisses => int.parse(
         _csgSync(
           'cacheMisses',
@@ -209,6 +209,8 @@ class StoreStats {
   /// Retrieve the number of tiles that were unsuccessfully retrieved from the store during browsing
   ///
   /// For synchronous version, see [cacheMisses].
+  ///
+  /// Includes tiles that needed updating. If using [noCache], this will always return 0.
   Future<int> get cacheMissesAsync async => int.parse(
         await _csgAsync(
           'cacheMisses',